Embed enum
